Evaluation fixes

+++ cpp/src/main.hpp
@@ -2,7 +2,7 @@
 
 void selfPlayMain(int runId, const std::string &savePath, int numProcessors, int numGPUs);
 
-typedef std::pair<int, std::vector<std::pair<int, int>>> PyInferenceResult; // (score, visits);
+typedef std::pair<float, std::vector<std::pair<int, int>>> PyInferenceResult; // (score, visits);
 
 std::vector<PyInferenceResult> boardInferenceMain(const std::string &modelPath,
                                                   const std::vector<std::string> &fens);
